+++ mobile/android/android-components/components/feature/contextmenu/src/test/java/mozilla/components/feature/contextmenu/ContextMenuFeatureTest.kt
@@ -9,8 +9,18 @@ import android.view.View
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.test.TestCoroutineDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.browser.state.action.ContentAction
+import mozilla.components.browser.state.action.TabListAction
+import mozilla.components.browser.state.selector.findTab
+import mozilla.components.browser.state.state.BrowserState
+import mozilla.components.browser.state.state.createTab
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.support.base.Component
@@ -18,42 +28,68 @@ import mozilla.components.support.base.facts.Action
 import mozilla.components.support.base.facts.processor.CollectionProcessor
 import mozilla.components.support.base.observer.Consumable
 import mozilla.components.support.test.any
+import mozilla.components.support.test.ext.joinBlocking
 import mozilla.components.support.test.mock
 import mozilla.components.support.test.robolectric.testContext
-import org.junit.Assert
+import org.junit.After
+import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
+import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.`when`
 import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.verify
-import org.mockito.Mockito.verifyNoMoreInteractions
-import java.util.UUID
 
 @RunWith(AndroidJUnit4::class)
 class ContextMenuFeatureTest {
+    private val testDispatcher = TestCoroutineDispatcher()
+
+    private lateinit var store: BrowserStore
+
+    @Before
+    fun setUp() {
+        Dispatchers.setMain(testDispatcher)
+
+        store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "test-tab")
+            ),
+            selectedTabId = "test-tab"
+        ))
+    }
+
+    @After
+    fun tearDown() {
+        Dispatchers.resetMain()
+        testDispatcher.cleanupTestCoroutines()
+    }
 
     @Test
     fun `New HitResult for selected session will cause fragment transaction`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply { sessionManager.add(this) }
-
         val fragmentManager = mockFragmentManager()
 
         val (engineView, view) = mockEngineView()
 
         val feature = ContextMenuFeature(
             fragmentManager,
-            sessionManager,
-            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock()),
-            engineView)
+            store,
+            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock(), mock()),
+            engineView,
+            mock())
 
         feature.start()
 
-        val hitResult = HitResult.UNKNOWN("https://www.mozilla.org")
-        session.hitResult = Consumable.from(hitResult)
+        store.dispatch(ContentAction.UpdateHitResultAction(
+            "test-tab",
+            HitResult.UNKNOWN("https://www.mozilla.org")
+        )).joinBlocking()
+
+        testDispatcher.advanceUntilIdle()
 
         verify(fragmentManager).beginTransaction()
         verify(view).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
@@ -61,24 +97,26 @@ class ContextMenuFeatureTest {
 
     @Test
     fun `New HitResult for selected session will not cause fragment transaction if feature is stopped`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply { sessionManager.add(this) }
-
         val fragmentManager = mockFragmentManager()
 
         val (engineView, view) = mockEngineView()
 
         val feature = ContextMenuFeature(
             fragmentManager,
-            sessionManager,
-            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock()),
-            engineView)
+            store,
+            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock(), mock()),
+            engineView,
+            mock())
 
         feature.start()
         feature.stop()
 
-        val hitResult = HitResult.UNKNOWN("https://www.mozilla.org")
-        session.hitResult = Consumable.from(hitResult)
+        store.dispatch(ContentAction.UpdateHitResultAction(
+            "test-tab",
+            HitResult.UNKNOWN("https://www.mozilla.org")
+        )).joinBlocking()
+
+        testDispatcher.advanceUntilIdle()
 
         verify(fragmentManager, never()).beginTransaction()
         verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
@@ -86,27 +124,29 @@ class ContextMenuFeatureTest {
 
     @Test
     fun `Feature will re-attach to already existing fragment`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply {
-            sessionManager.add(this)
-            hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
-        }
-
         val fragment: ContextMenuFragment = mock()
-        doReturn(session.id).`when`(fragment).sessionId
+        doReturn("test-tab").`when`(fragment).sessionId
 
         val fragmentManager: FragmentManager = mock()
         doReturn(fragment).`when`(fragmentManager).findFragmentByTag(any())
 
         val (engineView, view) = mockEngineView()
 
+        store.dispatch(ContentAction.UpdateHitResultAction(
+            "test-tab",
+            HitResult.UNKNOWN("https://www.mozilla.org")
+        )).joinBlocking()
+
         val feature = ContextMenuFeature(
             fragmentManager,
-            sessionManager,
-            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock()),
-            engineView)
+            store,
+            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock(), mock()),
+            engineView,
+            mock())
 
         feature.start()
+
+        testDispatcher.advanceUntilIdle()
 
         verify(fragment).feature = feature
         verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
@@ -114,13 +154,8 @@ class ContextMenuFeatureTest {
 
     @Test
     fun `Already existing fragment will be removed if session has no HitResult set anymore`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply {
-            sessionManager.add(this)
-        }
-
         val fragment: ContextMenuFragment = mock()
-        doReturn(session.id).`when`(fragment).sessionId
+        doReturn("test-tab").`when`(fragment).sessionId
 
         val transaction: FragmentTransaction = mock()
 
@@ -133,11 +168,14 @@ class ContextMenuFeatureTest {
 
         val feature = ContextMenuFeature(
             fragmentManager,
-            sessionManager,
-            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock()),
-            engineView)
+            store,
+            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock(), mock()),
+            engineView,
+            mock())
 
         feature.start()
+
+        testDispatcher.advanceUntilIdle()
 
         verify(fragmentManager).beginTransaction()
         verify(transaction).remove(fragment)
@@ -146,10 +184,8 @@ class ContextMenuFeatureTest {
     }
 
     fun `Already existing fragment will be removed if session does not exist anymore`() {
-        val sessionManager = SessionManager(mock())
-
         val fragment: ContextMenuFragment = mock()
-        doReturn(UUID.randomUUID().toString()).`when`(fragment).sessionId
+        doReturn("test-tab").`when`(fragment).sessionId
 
         val transaction: FragmentTransaction = mock()
 
@@ -162,11 +198,17 @@ class ContextMenuFeatureTest {
 
         val feature = ContextMenuFeature(
             fragmentManager,
-            sessionManager,
-            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock()),
-            engineView)
+            store,
+            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock(), mock()),
+            engineView,
+            mock())
+
+        store.dispatch(TabListAction.RemoveTabAction("test-tab"))
+            .joinBlocking()
 
         feature.start()
+
+        testDispatcher.advanceUntilIdle()
 
         verify(fragmentManager).beginTransaction()
         verify(transaction).remove(fragment)
@@ -188,49 +230,61 @@ class ContextMenuFeatureTest {
         val (engineView, view) = mockEngineView()
 
         val feature = ContextMenuFeature(
-            fragmentManager, mock(), listOf(candidate), engineView)
+            fragmentManager,
+            store,
+            listOf(candidate),
+            engineView,
+            ContextMenuUseCases(mock())
+        )
 
-        feature.onLongPress(
-            Session("https://www.mozilla.org"),
+        feature.showContextMenu(
+            createTab("https://www.mozilla.org"),
             HitResult.UNKNOWN("https://www.mozilla.org"))
 
-        verifyNoMoreInteractions(fragmentManager)
+        verify(fragmentManager, never()).beginTransaction()
         verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
-    fun `Selecting context menu item will consume HitResult`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply {
-            sessionManager.add(this)
-            hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
+    fun `Cancelling context menu item will consume HitResult`() {
+        store = BrowserStore()
+
+        val sessionManager = SessionManager(engine = mock(), store = store)
+        Session("https://www.mozilla.org", id = "test-tab").also {
+            sessionManager.add(it)
+            it.hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
         }
 
-        val (engineView, view) = mockEngineView()
+        val (engineView, _) = mockEngineView()
 
         val feature = ContextMenuFeature(
-            mock(),
-            sessionManager,
-            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock()),
-            engineView)
+            mockFragmentManager(),
+            store,
+            ContextMenuCandidate.defaultCandidates(testContext, mock(), mock(), mock()),
+            engineView,
+            ContextMenuUseCases(sessionManager)
+        )
 
-        assertFalse(session.hitResult.isConsumed())
+        assertNotNull(store.state.findTab("test-tab")!!.content.hitResult)
 
-        feature.onMenuCancelled(session.id)
+        feature.onMenuCancelled("test-tab")
 
-        assertTrue(session.hitResult.isConsumed())
+        testDispatcher.advanceUntilIdle()
 
-        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
+        assertNull(store.state.findTab("test-tab")!!.content.hitResult)
     }
 
     @Test
     fun `Selecting context menu item will invoke action of candidate and consume HitResult`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply {
-            sessionManager.add(this)
-            hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
+        store = BrowserStore()
+
+        val sessionManager = SessionManager(engine = mock(), store = store)
+        Session("https://www.mozilla.org", id = "test-tab").also {
+            sessionManager.add(it)
+            it.hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
         }
 
+        val (engineView, view) = mockEngineView()
         var actionInvoked = false
 
         val candidate = ContextMenuCandidate(
@@ -239,57 +293,63 @@ class ContextMenuFeatureTest {
             showFor = { _, _ -> true },
             action = { _, _ -> actionInvoked = true })
 
-        val (engineView, view) = mockEngineView()
-
         val feature = ContextMenuFeature(
-            mock(),
-            sessionManager,
+            mockFragmentManager(),
+            store,
             listOf(candidate),
-            engineView)
+            engineView,
+            ContextMenuUseCases(sessionManager)
+        )
 
-        assertFalse(session.hitResult.isConsumed())
+        testDispatcher.advanceUntilIdle()
+
+        assertNotNull(store.state.findTab("test-tab")!!.content.hitResult)
         assertFalse(actionInvoked)
 
-        feature.onMenuItemSelected(session.id, candidate.id)
+        feature.onMenuItemSelected("test-tab", "test-id")
 
-        assertTrue(session.hitResult.isConsumed())
+        testDispatcher.advanceUntilIdle()
+
+        assertNull(store.state.findTab("test-tab")!!.content.hitResult)
         assertTrue(actionInvoked)
-
         verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
     fun `Selecting context menu item will emit a click fact`() {
-        val sessionManager = SessionManager(mock())
-        val session = Session("https://www.mozilla.org").apply {
-            sessionManager.add(this)
-            hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
+        store = BrowserStore()
+
+        val sessionManager = SessionManager(engine = mock(), store = store)
+        Session("https://www.mozilla.org", id = "test-tab").also {
+            sessionManager.add(it)
+            it.hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
         }
 
-        val candidate = ContextMenuCandidate(
-                id = "test-id",
-                label = "Test Item",
-                showFor = { _, _ -> true },
-                action = { _, _ -> /* noop */ })
-
         val (engineView, _) = mockEngineView()
+        val candidate = ContextMenuCandidate(
+            id = "test-id",
+            label = "Test Item",
+            showFor = { _, _ -> true },
+            action = { _, _ -> /* noop */ })
 
         val feature = ContextMenuFeature(
-                mock(),
-                sessionManager,
-                listOf(candidate),
-                engineView)
+            mockFragmentManager(),
+            store,
+            listOf(candidate),
+            engineView,
+            ContextMenuUseCases(sessionManager)
+        )
 
         CollectionProcessor.withFactCollection { facts ->
-            feature.onMenuItemSelected(session.id, candidate.id)
+            feature.onMenuItemSelected("test-tab", candidate.id)
 
-            Assert.assertEquals(1, facts.size)
+            assertEquals(1, facts.size)
 
             val fact = facts[0]
-            Assert.assertEquals(Component.FEATURE_CONTEXTMENU, fact.component)
-            Assert.assertEquals(Action.CLICK, fact.action)
-            Assert.assertEquals("item", fact.item)
-            Assert.assertEquals("test-id", fact.metadata?.get("item"))
+            assertEquals(Component.FEATURE_CONTEXTMENU, fact.component)
+            assertEquals(Action.CLICK, fact.action)
+            assertEquals("item", fact.item)
+            assertEquals("test-id", fact.metadata?.get("item"))
         }
     }